feat

+++ Proyecto_ED_Grupo1/src/CampoBatalla.java
@@ -11,7 +11,7 @@ public class CampoBatalla {
     
     public void campoBatalla(){
         long startTime = System.currentTimeMillis();
-        int vidaCastCPU=10, vidaCastJug=10;
+        double vidaCastCPU=10, vidaCastJug=10;
         String[][] matriz=new String[3][8];
         for (int i=0;i<3;i++){
             for (int j=0;j<8;j++){
@@ -22,8 +22,8 @@ public class CampoBatalla {
         matriz[1][1]="C";
         matriz[1][6]="C";
         matriz[0][2]="M";
-        matriz[1][0]=Integer.toString(vidaCastJug);
-        matriz[1][7]=Integer.toString(vidaCastCPU);
+        matriz[1][0]=Double.toString(vidaCastJug);
+        matriz[1][7]=Double.toString(vidaCastCPU);
         for (int i=0;i<3;i++){
                 for (int j=0;j<8;j++){
                    System.out.print(matriz[i][j]+" "); 
@@ -32,8 +32,8 @@ public class CampoBatalla {
         } 
         vidaCastJug-=2;
         vidaCastCPU-=2;
-        matriz[1][0]=Integer.toString(vidaCastJug);
-        matriz[1][7]=Integer.toString(vidaCastCPU);
+        matriz[1][0]=Double.toString(vidaCastJug);
+        matriz[1][7]=Double.toString(vidaCastCPU);
         matriz[0][2]="-";
         matriz[0][3]="M";
         for (int i=0;i<3;i++){
